+++ src/oapv_tbl.h
@@ -36,7 +36,7 @@
 
 extern const u8  oapv_tbl_log2[257];
 extern const s8  oapv_tbl_tm8[8][8];
-extern const int oapv_tbl_dq_scale[6];
+extern const u8  oapv_tbl_dq_scale[6];
 extern const u8  oapv_tbl_scan[OAPV_BLK_D];
 extern int       oapv_itrans_diff[64][64];
 extern const u16 oapve_tbl_vlc_code[100][5][2];
